refact: refatoração do segundo teste e aplicação da verificação dos dados do componente CardForm para poder habilitar o botão de save
